#53 [refactor] : 캔들 유틸로 분리
